+++ changelog.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### enhancements
 - store service logs in a file to help debugging.
+- wait for native shell on startup before start seelen ui.
 
 ## [2.2.7]
 ### fix